feat: Adding ContentDialog and MessageDialogs

+++ src/ToDo/Presentation/TaskListViewModel.cs
@@ -34,13 +34,26 @@ public partial class TaskListViewModel: IRecipient<EntityMessage<ToDoTask>>
 	}
 
 	// TODO: Feed - This should be a ListFeed / This should listen for Task creation/update/deletion
-	public IFeed<IImmutableList<ToDoTask>> Tasks => _entity.SelectAsync(async (list, ct) => await _listSvc.GetTasksAsync(list!, ct));
+	public IFeed<IImmutableList<ToDoTask>> Tasks => _entity.SelectAsync(async (list, ct) => list is not null ? await _listSvc.GetTasksAsync(list, ct): default);
 
 	private async ValueTask CreateTask(ToDoTaskList list, CancellationToken ct)
 	{
-		// TODO: Configure properties of TaskData
-		var newTask = new ToDoTask {Title = "Hello world"};
-		await _taskSvc.CreateAsync(list, newTask, ct);
+		var response = await _navigator!.NavigateViewModelForResultAsync<AddTaskViewModel, ToDoTaskData>(this, qualifier: Qualifiers.Dialog);
+		if (response is null)
+		{
+			return;
+		}
+
+		var result = await response.Result;
+
+		var taskName = result.SomeOrDefault()?.Title;
+		if (taskName is not null)
+		{
+
+			// TODO: Configure properties of TaskData
+			var newTask = new ToDoTask { Title = taskName };
+			await _taskSvc.CreateAsync(list, newTask, ct);
+		}
 	}
 
 	private async ValueTask NavigateToTask(ToDoTask task, CancellationToken ct)
@@ -51,8 +64,19 @@ public partial class TaskListViewModel: IRecipient<EntityMessage<ToDoTask>>
 
 	private async ValueTask DeleteList(ToDoTaskList list, CancellationToken ct)
 	{
-		await _listSvc.DeleteAsync(list, ct);
-		await _navigator.NavigateBackAsync(this, cancellation: ct);
+		var response = await _navigator!.NavigateRouteForResultAsync<DialogAction>(this, "Confirm", qualifier: Qualifiers.Dialog);
+		if (response is null)
+		{
+			return;
+		}
+
+		var result = await response.Result;
+		if (result.SomeOrDefault()?.Id?.ToString() == "Y")
+		{
+
+			await _listSvc.DeleteAsync(list, ct);
+			await _navigator.NavigateBackAsync(this, cancellation: ct);
+		}
 	}
 
 	public void Receive(EntityMessage<ToDoTask> msg)
