+++ src/ToDo/Presentation/TaskViewModel.cs
@@ -21,8 +21,18 @@ public partial class TaskViewModel
 
 	private async ValueTask Delete(ToDoTask task, CancellationToken ct)
 	{
-		await _svc.DeleteAsync(task, ct);
-		await _navigator.NavigateBackAsync(this, cancellation: ct);
+		var response = await _navigator!.NavigateRouteForResultAsync<DialogAction>(this, "Confirm", qualifier: Qualifiers.Dialog);
+		if (response is null)
+		{
+			return;
+		}
+
+		var result = await response.Result;
+		if (result.SomeOrDefault()?.Id?.ToString() == "Y")
+		{
+			await _svc.DeleteAsync(task, ct);
+			await _navigator.NavigateBackAsync(this, cancellation: ct);
+		}
 	}
 
 	private async ValueTask Save(ToDoTask task, CancellationToken ct)
